Refactored tests folder & files structure
