+++ tests/integration/vendors/test_rabbitmq.py
@@ -3,6 +3,7 @@ from kombu import Connection
 from pytest_lazyfixture import lazy_fixture
 
 from pytest_celery import RabbitMQContainer
+from pytest_celery import RabbitMQTestBroker
 from pytest_celery import defaults
 
 
@@ -15,3 +16,9 @@ class test_rabbitmq_container:
             assert c.connect()
         finally:
             c.release()
+
+
+@pytest.mark.parametrize("node", [lazy_fixture(defaults.CELERY_RABBITMQ_BROKER)])
+class test_rabbitmq_test_broker:
+    def test_ready(self, node: RabbitMQTestBroker):
+        assert node.ready()
