+++ tests/unit/vendors/test_redis.py
@@ -1,4 +1,6 @@
 from pytest_celery import RedisContainer
+from pytest_celery import RedisTestBackend
+from pytest_celery import RedisTestBroker
 from pytest_celery import defaults
 
 
@@ -18,3 +20,13 @@ class test_redis_container:
 
     def test_image(self, redis_test_container: RedisContainer):
         assert redis_test_container.image() == defaults.REDIS_IMAGE
+
+
+class test_redis_test_backend:
+    def test_ready(self, celery_redis_backend: RedisTestBackend):
+        assert celery_redis_backend.ready()
+
+
+class test_redis_test_broker:
+    def test_ready(self, celery_redis_broker: RedisTestBroker):
+        assert celery_redis_broker.ready()
